Replaces sidebar with tab navigation and refactors a few components to make them more reusable (#6)
New logo! But wait... that's not all. Glad to finally have the new branding finished but I've also basically refactored the whole app. The actual tool now lives as an SPA after logging in. Thanks to NextJS I've been able to carve the app up into different routes that run best depending on the content. Some are rendered serverside while others client side.

It's been a long time coming but Hemolog is back! This commit means I'm ready to launch.