Mobile design improvements (#12)
There were a number of issues with the header and footer on mobile. I also took this time to add a social image and clean up a few bits of layout that were bugging me on desktop.

The largest change here is the addition of /changelog. The Changelog is our development blog that will chronical updates and changes to Hemolog going forward.